backend/requirements.txt
# Web framework
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0

# Database
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0
redis>=5.0,<6.0

# Data sources
yfinance>=0.2.40
FinMind>=1.6
tqdm>=4.66

# Numeric computation (pulled in transitively by yfinance; pinned explicitly)
numpy>=1.26,<3.0
pandas>=2.2,<3.0

# Validation & serialization
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
email-validator>=2.0,<3.0

# Auth
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=4.0,<5.0

# HTTP client
httpx>=0.27,<1.0

# Task queue
celery[redis]>=5.4,<6.0

# Dev & testing
pytest>=8.2,<9.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=5.0,<6.0
fakeredis[aioredis]>=2.21
aiosqlite>=0.20
black>=24.4,<25.0
ruff>=0.4,<1.0
